==> denguePaper/analysis/02_mosquito-processing.Rmd <==
---
title: "Cleaning and Processing Mosquito Data"
author: "Michelle Evans"
date: ""
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = T)
knitr::opts_chunk$set(cache=TRUE)
knitr::opts_chunk$set(message=FALSE)
knitr::opts_chunk$set(warning=FALSE)

#packages
library(dplyr)
library(tidyr)
```

This document contains the code for the cleaning and processing of mosquito emergence and infection data in 'Carry-over effects of larval microclimate on the transmission potential of a mosquito-borne pathogen' (Evans et al. 2018).

Paper Citation:

Evans MV, Shiau JC, Solano N, Brindley MA, Drake JM, Murdock, CC. 2018. Carry-over effects of larval microclimate on the transmission potential of a mosquito-borne pathogen.  

# Infection Data

```{r formatting function}
formatData <- function(month){
  #' format infection data
  #' @params month (ie. "august")
  #' @returns dataframe of properly formatted data
  #adjust wingLength
  monthDf <- read.csv(paste0("denguePaper/data/infections/raw/", month,"Dengue.csv"))
  #convert wingLength and drop extra columns
  monthDf$Wing <- monthDf$WingLength*monthDf$conversion..mm.bar.
  monthDf <- dplyr::select(monthDf, -WingLength, -conversion..mm.bar.)
  
  #dpi as factor
  monthDf$DPI <- as.factor(monthDf$DPI)
  
  #add in class and site
  monthDf$site <- as.factor(substr(as.character(monthDf$Individual), 1, 2))
  monthDf$class <- NULL
  for (i in 1:nrow(monthDf)){
    if (substr(monthDf$site[i], 1,1)=="R"){
    monthDf$class[i] <- "Rural"
    } else if (substr(monthDf$site[i], 1,1)=="S"){
    monthDf$class[i] <- "Suburban"
    } else if (substr(monthDf$site[i], 1,1)=="U"){
    monthDf$class[i] <- "Urban"
    }
  }
  monthDf$class <- as.factor(monthDf$class)
  
  #convert Y and N to 1 and 0 for statistics
  levels(monthDf$Body) <- c("NA", 0, 1)
  monthDf$Body <- as.numeric(as.character(monthDf$Body))
  levels(monthDf$Saliva) <- c("NA", 0, 1)
  monthDf$Saliva <- as.numeric(as.character(monthDf$Saliva))
  # august had no contaminated heads, so different corrections
  if (month=="august"){
    levels(monthDf$Head) <- c(0, 1)
  } else levels(monthDf$Head) <- c("NA",0, 1)
  monthDf$Head <- as.numeric(as.character(monthDf$Head))
  
  ##Fix false negatives
  #adjust so that if saliva is positive, so is head
  #ddjust so that is head is positive, so is body
  monthDf$Head[monthDf$Saliva>0] <- 1
  monthDf$Body[monthDf$Head>0] <- 1
  
  return(monthDf)
}
```

Format each month via above function and combine together.

```{r}
summerInfection <- formatData("august")
fallInfection <- formatData("october")
seasonInfection <- rbind(summerInfection, fallInfection)
seasonInfection$Block <- factor(c(rep("Summer", nrow(summerInfection)), rep("Fall", nrow(fallInfection))), levels = c("Summer", "Fall"))

seasonInfection <- seasonInfection %>%
  rename(Site_ID = site, Class = class) %>%
  #only need 21 dpi
  filter(DPI==21) %>%
  select(-DPI)
```

##  Summary

Calculate the mean and sd for plots per block and class.

```{r}
seasonInfSummary <- seasonInfection %>%
  #drop individual
  dplyr::select(-Individual, -Site_ID, -Wing) %>%
  group_by(Block, Class) %>%
  summarise_all(funs(mean(.,na.rm=T),sd(.,na.rm=T),se=(sd(., na.rm=T)/sqrt(n())))) %>%
  ungroup()
```

```{r}
saveRDS(seasonInfection, "denguePaper/data/infections/clean/seasonInfection.RData")

saveRDS(seasonInfSummary, "denguePaper/data/infections/clean/infectionSummary.RData")
```

## Long Infection Data

This data is long for the plot.

```{r}
infLong <- seasonInfection %>%
  gather(type, infection, Body:Saliva) %>%
  dplyr::select(Block, Class, Site_ID, Individual, type, infection) %>%
  dplyr::group_by(Block, Class, type, Site_ID) %>%
  summarise(mean.inf = mean(infection, na.rm=T), 
            sampleSize = sum(!is.na(infection)), 
            positive = sum(infection, na.rm=T)) %>%
  group_by(Block, Class, type) %>%
  summarise(se.inf = sd(mean.inf, na.rm=T)/n(), 
            mean.inf = mean(mean.inf), 
            samples = sum(sampleSize), 
            positives = sum(positive)) %>%
  ungroup() %>%
  mutate(stripLabel=case_when(
    type=="Body" ~ "Infected",
    type=="Head" ~ "Disseminated",
    type=="Saliva" ~ "Infectious"
  ))
infLong$sampleLab <- paste0(infLong$positives, "(", infLong$samples, ")")
```

```{r}
saveRDS(infLong, "denguePaper/data/infections/clean/infectionSummaryLong.RData")
```

## Infection Efficiency

```{r}
bodyEff <- seasonInfection %>%
  dplyr::select(-Individual, -Site_ID, -Wing, -Head, -Saliva) %>%
  group_by(Block, Class) %>%
  summarise(bodyMean=mean(Body,na.rm=T), bodySE=(sd(Body, na.rm=T)/sqrt(n()))) %>%
  ungroup()

headEff <- seasonInfection %>%
  filter(Body==1) %>%
  #drop individual
  dplyr::select(-Individual, -Site_ID, -Wing,  -Body, -Saliva) %>%
  group_by(Block, Class) %>%
  summarise(headMean=mean(Head,na.rm=T),headSE=(sd(Head, na.rm=T)/sqrt(n()))) %>%
  ungroup()

salEff <-  seasonInfection %>%
  filter(Head==1) %>%
  #drop individual
  dplyr::select(-Individual, -Site_ID, -Wing, -Head, -Body) %>%
  group_by(Block, Class) %>%
  summarise(salMean=mean(Saliva,na.rm=T),salSE=(sd(Saliva, na.rm=T)/sqrt(n()))) %>%
  ungroup()

#group together
allEff <- full_join(bodyEff, headEff, by=c("Block", "Class"))
allEff <- full_join(allEff, salEff, by=c("Block", "Class"))
```

For plotting
```{r}
meltMean <- allEff %>%
  select(Block, Class, contains("Mean")) %>%
  gather(key=variable, value=mean, -Block, -Class) 

meltMean$type <- rep(c("Body", "Head", "Saliva"), each=6)
  
meltSE <- allEff %>%
  select(Block, Class, contains("SE")) %>%
  gather(variable, SE, -Block, -Class)

meltSE$type <- rep(c("Body", "Head", "Saliva"), each=6)
  
meltAll <- full_join(meltMean, meltSE, by=c("Class", "Block", "type")) %>%
  select(-variable.x, -variable.y)
```

```{r}
saveRDS(meltAll, "denguePaper/data/infections/clean/efficiencyPlot.RData")
```

# Emergence Data

```{r}
augEmerg <- read.csv("denguePaper/data/emergence/raw/AugustEmergence.csv")
augEmerg$Block <- as.factor("Summer")
augEmerg$Class <- factor(Hmisc::capitalize(as.character(augEmerg$Class)), levels = c("Rural", "Suburban", "Urban"))
augEmerg <- rename(augEmerg, Site_ID = Site_Code, Tray_ID = Tray_Code)

octEmerg <- read.csv("denguePaper/data/emergence/raw/OctoberEmergence.csv")
octEmerg$Block <- as.factor("Fall")
octEmerg$Class <- factor(Hmisc::capitalize(as.character(octEmerg$Class)), levels = c("Rural", "Suburban", "Urban"))
octEmerg <- rename(octEmerg, Site_ID = Site_Code, Tray_ID = Tray_Code)

#drop trays that were disturbed by wildlife
octEmerg <- filter(octEmerg, Tray_ID!="U3T1") #ants invaded and killed mosquitoes
octEmerg <- filter(octEmerg, Tray_ID!="S1T3") #dumped
octEmerg <- filter(octEmerg, Tray_ID!="U1T4") #dumped

allEmerg <- rbind(augEmerg, octEmerg)
```

Note that two trays were true zeros in the fall replicate: R2T4 and U3T4

Expand so that each mosquito is on its own row.

```{r}
emergExp <- allEmerg[rep(seq.int(1,nrow(allEmerg)), allEmerg$Num_Emerge),
                      c(colnames(allEmerg) %in% c('Block', 'Class', 'Site_ID', 'Tray_ID','Exp_Day', 'Sex'))]
#sum(allEmerg$Num_Emerge)==nrow(emergExp) #quick check this worked
```

## Development Rate

```{r}
emergTray <- emergExp %>%
  filter(Sex=="F") %>%
  mutate(devRate=1/Exp_Day) %>%
  group_by(Tray_ID, Site_ID, Class, Block) %>%
  summarise(devRate=mean(devRate)) %>%
  ungroup() %>%
  #add true zeros back in
  bind_rows(data.frame(Block = "Fall", 
                       Site_ID = c("R2", "U3"), 
                       Class = c("Rural", "Urban"), 
                       Tray_ID = c("R2T4", "U3T4"),
                       devRate = 0))

#add factors back in
emergTray$Tray_ID <- as.factor(emergTray$Tray_ID)
emergTray$Site_ID <- as.factor(emergTray$Site_ID)
emergTray$Class <- as.factor(emergTray$Class)
emergTray$Block <- factor(emergTray$Block, levels = c("Summer", "Fall"))
```

## Survival

```{r}
#get survival per tray
survSumm <- emergExp %>%
  filter(Sex=="F") %>%
  group_by(Block, Tray_ID) %>%
  dplyr::mutate(numFSurv=n()) %>% #does same as summarise but keeps factor labels
  ungroup() %>%
  dplyr::select(Block, Class, Site_ID, Tray_ID, numFSurv) %>%
  distinct() %>% #drop duplicates
  #add true zeros back in
  bind_rows(data.frame(Block = "Fall", 
                       Site_ID = c("R2", "U3"), 
                       Class = c("Rural", "Urban"), 
                       Tray_ID = c("R2T4", "U3T4"),
                       numFSurv = 0))

#add factors back in
survSumm$Tray_ID <- as.factor(survSumm$Tray_ID)
survSumm$Site_ID <- as.factor(survSumm$Site_ID)
survSumm$Class <- as.factor(survSumm$Class)
survSumm$Block <- factor(survSumm$Block, levels = c("Summer", "Fall"))

```

## Individual Level Measurements

### Fill-in Functions

```{r fillIn function}
fillIn <- function(df, endDay, totalMosq=50){
  #' Fill In Emergence Dates
  #' this function fills in for those mosquitoes that did not emerge so we do not have data for, it gives them an observation/event of 0 on the last day we found a mosquito emerged
  #' @param df the data frame you wish to fill in, in our case by pot
  #' @param endDay the last day of emergence
  #' @param totalMosq estimated starting number of mosquitoes per pot
  #' @returns dataframe with census data filled in for mosquitoes that did not emerge

  toRep <- df[1,]
  toRep$Exp_Day <- endDay
  toRep$event <- 0
  if(nrow(df)<totalMosq){
    toAdd <- toRep[rep(1, (totalMosq-nrow(df))),]
    allTest <- rbind(df, toAdd)
  } else {
    toAdd <- NA
    allTest <- NA
  }
  return(allTest)
}
```

```{r apply FillIn function}
applyFill <- function(season, allData=emergExp){
  #' Apply FillIn function
  #' @param season "fall" or "summer"
  #' @param allData full dataframe with row for each mosquito that emerged
  #' @returns censused data for the full season
  
  tempList <- list()
  tempDF <- allData
  tempDF <- tempDF[tempDF$Sex=="F",]
  tempDF <- tempDF[tempDF$Block==season,]
  tempDF$event <- 1 #add emergence event
  endDay <- max(tempDF$Exp_Day)
  for (i in 1:length(levels(tempDF$Tray_ID))){ 
    df <- tempDF[tempDF$Tray_ID==levels(tempDF$Tray_ID)[i],]
    if (nrow(df)==0){
      #if no mosquitoes emerged for that tray, create a zero emergence
      df <- allEmerg%>%
        select(Site_ID, Tray_ID, Class, Exp_Day, Sex, Block) %>%
        filter(Block == season & Tray_ID == levels(tempDF$Tray_ID)[i]) %>%
        slice(1) %>%
        mutate(Sex = factor("F", levels = c("F", "M"))) %>%
        mutate(event = 0) %>%
        mutate(Exp_Day = endDay)
    }
    #skip failed trays in fall
    if (season == "Fall" & df$Tray_ID[1] %in% c("U3T1", "S1T3", "U1T4")){
      next
    } else {
    tempList[[i]] <- fillIn(df=df, endDay=endDay)
    }
  }
  allSurv <- do.call(rbind.data.frame, tempList)
  allSurv <- na.omit(allSurv)
  return(allSurv)
}
```

### Survival


```{r}
summerSurv <- applyFill(season="Summer")
fallSurv <- applyFill(season="Fall")
allSurv <- rbind(summerSurv, fallSurv)
```

## Save

```{r}
saveRDS(allSurv, "denguePaper/data/emergence/clean/individuals.RData")
saveRDS(emergTray, "denguePaper/data/emergence/clean/emergenceTray.RData")
saveRDS(survSumm, "denguePaper/data/emergence/clean/survivalTray.RData")
```

# Wing Length

## Uninfected Body Size

Load body size data:
```{r load body size data}
augWing <- read.csv("denguePaper/data/emergence/raw/AugustWingLength.csv", stringsAsFactors = F)
octWing <- read.csv("denguePaper/data/emergence/raw/OctoberWingLength.csv",  stringsAsFactors = F)


#convert to mm & clean
augWing$mm <- augWing$Bars*augWing$Conversion.mm.bars.
octWing$mm <- octWing$Bars*octWing$Conversion.bars.mm.

octWing$Site_ID <- as.factor(substr(as.character(octWing$TrayCode), 1, 2))
augWing$Site_ID <- as.factor(substr(as.character(augWing$TrayCode), 1, 2))

getClass <- function(monthDf){
  monthDf$Class <- NULL
  for (i in 1:nrow(monthDf)){
    if (substr(monthDf$Site_ID[i], 1,1)=="R"){
    monthDf$Class[i] <- "Rural"
    }
    if (substr(monthDf$Site_ID[i], 1,1)=="S"){
    monthDf$Class[i] <- "Suburban"
    }
    if (substr(monthDf$Site_ID[i], 1,1)=="U"){
    monthDf$Class[i] <- "Urban"
    }
  }
  monthDf$Class <- factor(monthDf$Class, levels = c("Rural", "Suburban", "Urban"))
  return(monthDf)
}

augWing <- getClass(augWing)
octWing <- getClass(octWing)

octWing$Block <- "Fall"
augWing$Block <- "Summer"

augWing$Date <- as.Date(as.character(augWing$Date), format="%m/%d/%Y")
octWing$Date <- as.Date(as.character(octWing$Date), format="%m/%d/%Y")

#add day of experiment
augWing$Exp_Day <- as.numeric(augWing$Date-as.Date("2016-08-01", format="%Y-%m-%d"))
octWing$Exp_Day <- as.numeric(octWing$Date-as.Date("2016-09-26", format="%Y-%m-%d"))

#combine
fallWing <- octWing %>%
  dplyr::select(Block, Class, Site_ID, Tray_ID = TrayCode, Exp_Day, Date, mm)

summerWing <- augWing %>%
  dplyr::select(Block, Class, Site_ID, Tray_ID = TrayCode, Exp_Day, Date, mm)

allWing <- rbind(summerWing, fallWing)
allWing$Block<- factor(allWing$Block, levels=c("Summer", "Fall"))
allWing$Tray_ID <- factor(allWing$Tray_ID)

#drop outlier in S2 (wing size =1.56 mm)
allWing <- allWing %>%
  filter(mm>1.6)
```

## Add Climate Data to Wing Length

This connects climate data for each individual mosquito, from the start day of each replicate until the day it emerged

```{r}
climate <- readRDS("denguePaper/data/microclimate/clean/2016TrialsAdult.RData")
```

```{r}
getClimate <- function(indMosq, climateDF=climate){
  #' This is a function to apply over the rows of the octWing and augWing data frames. Must have climate data loaded
  
  #' @param indMosq row of the dataframe for each individual mosquito
  #' @param climateDF the dataframe containing climate data every 10 minutes
  #' @returns formatted data with climate and winglength for the individual mosquito
  
  
  #get date range
  startDate <- ifelse(indMosq$Block=="Summer", "2016-08-01", "2016-09-26")
  startDate <- as.Date(startDate, format="%Y-%m-%d")
  endDate <- indMosq$Date
  
  #subset temperature data
  try(climSubset <- climateDF %>%
    filter(Tray_ID==indMosq$Tray_ID) %>%
    filter(Day>startDate & Day<endDate),
    silent=T)
  
  #now take mean temperature
  tempMean <- climSubset %>%
    summarise(Tmean=mean(Temp, na.rm=T))
  
  return(tempMean$Tmean)
}
```

Note that some will have `NA` because the data logger failed in that tray.

```{r}
allWing$Temp <- NA
for(i in 1:nrow(allWing)){
  indMosq <- allWing[i,]
  allWing$Temp[i] <- getClimate(indMosq)
}
```

## Save

```{r}
saveRDS(allWing, "denguePaper/data/emergence/clean/wingLength.RData")
```
