==> denguePaper/analysis/03-model-calculations.Rmd <==
---
title: "Calculations for Growth and VC Models"
author: "Michelle Evans"
date: ""
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#packages
library(dplyr)
```

This document calculates the growth rate and VC based on field-derived and literature based data.

# Growth Rate

## Load Data

```{r}
allWing <- readRDS("denguePaper/data/emergence/clean/wingLength.RData")
emergence <- readRDS("denguePaper/data/emergence/clean/individuals.RData")
climateTray <- readRDS("denguePaper/data/microclimate/clean/climateTraySummary.RData")
```

## Calculate Tray Level Growth Rates

Data cleaning, mainly imputing for missing values.

```{r}
#get mean winglenth per tray and day of emergence
wingTray <- allWing %>%
  dplyr::group_by(Block, Class, Site_ID, Tray_ID, Exp_Day) %>%
  dplyr::summarise(meanWing=mean(mm, na.rm=T)) %>%
  dplyr::ungroup()

#subset out emerging females
emergF <- emergence %>%
  filter(Sex == "F" & event == 1)

#merge with emergence rates
growthDF <- merge(emergF, wingTray, by = c("Block", "Class", "Site_ID", "Tray_ID", "Exp_Day"), all.x = T)

#females that were used in infections may not have winglengths, they receive the tray level mean for that block
trayMeans <- allWing %>%
  dplyr::group_by(Block, Tray_ID) %>%
  dplyr::summarise(meanWing=mean(mm, na.rm=T))

#fill in missing with mean of tray during that Block
for (i in 1:nrow(growthDF)){
  if (is.na(growthDF$meanWing[i])){
    temp <- trayMeans$meanWing[trayMeans$Block==growthDF$Block[i] & trayMeans$Tray_ID==growthDF$Tray_ID[i]]
    if (length(temp)==0) next
    growthDF$meanWing[i] <- temp
  } else next
}

rm(trayMeans)

#some trays had mosquitoes that were only used for infections, they receive the site level mean to calculate their growth rates
siteMeans <- allWing %>%
  dplyr::group_by(Site_ID, Block) %>%
  dplyr::summarise(meanWing=mean(mm, na.rm=T))

for (i in 1:nrow(growthDF)){
  if (is.na(growthDF$meanWing[i])){
    temp <- siteMeans$meanWing[siteMeans$Block==growthDF$Block[i] & siteMeans$Site_ID==growthDF$Site_ID[i]]
    if (length(temp)==0) next
    growthDF$meanWing[i] <- temp
  } else next
}

rm(siteMeans)

growthDF <- growthDF %>%
  group_by(Block, Class, Site_ID, Tray_ID, Exp_Day) %>%
  summarise(Ax = sum(event), meanWing = mean(meanWing)) %>%
  ungroup()

```

Calculate growth rate :

```{r}
growthDF$Fwx <- -121.240 + (78.02 * growthDF$meanWing)
growthDF$AxFwx <- growthDF$Ax*growthDF$Fwx
growthDF$xAxFwx <- growthDF$Exp_Day*growthDF$AxFwx

#get sum per day
growthDF2 <- growthDF %>%
  dplyr::group_by(Block, Tray_ID, Class, Site_ID) %>%
  dplyr::summarise(xAxFwx = sum(xAxFwx), AxFwx=sum(AxFwx))

growthDF2 <- growthDF2 %>%
  mutate(r=(log((1/50)*AxFwx))/(14+(xAxFwx/AxFwx))) %>%
  ungroup() 
```

Add in temperature data:

```{r}
growthTemp <- growthDF2 %>%
  select(Block, Class, Site_ID, Tray_ID, r) %>%
  left_join(climateTray, by = c("Block", "Class", "Site_ID","Tray_ID"))
```

## Save

```{r}
saveRDS(growthTemp, "denguePaper/data/emergence/clean/growthRates.RData")
```

# Vectorial Capacity

## Load Data

```{r}
allWing <- readRDS("denguePaper/data/emergence/clean/wingLength.RData")
emergeTray <- readRDS("denguePaper/data/emergence/clean/emergenceTray.RData")
survTray <- readRDS("denguePaper/data/emergence/clean/survivalTray.RData")
infInd <- readRDS("denguePaper/data/infections/clean/seasonInfection.RData")
```

## Calculate Values at Constant 27C

These are adult traits that we didn't measure. The mosquitoes were kept at a constant 27C, so this is what we calculate them at. These are calculated at the site level because that is what level the infection data is at.

Briere:  $y = a * x * (x - t_0) * (t_{max} - x)^{(1/2)}$
Quad: $y = a * (x-t_0) * (x-t_{max})$

```{r}
parameters <- survTray %>%
  select(-numFSurv, -Tray_ID) %>%
  mutate(a = (1.93/10000)*27*(27-10.25)*((38.32-27)^0.5)) %>%
  mutate(PDR = (1.09/10000)*27*(27-10.39)*((43.05-27)^0.5)) %>%
  mutate(lf = -1.43*(27- 13.41)*(27-31.51)) %>%
  mutate(mu = 1/lf) %>%
  distinct()

``` 

## Calculate Field Derived Parameters

These are not carry-over effects, but larval survival (pEA) and emergence rate (MDR).

```{r}
pEA <- survTray %>%
  group_by(Block, Class, Site_ID) %>%
  dplyr::summarise(pEA=mean(numFSurv/50)) %>%
  ungroup()

MDR <- emergeTray %>%
  group_by(Block, Class, Site_ID) %>%
  dplyr::summarise(MDR=mean(devRate)) %>%
  ungroup()
```

## Calculate Carry-over Effect Parameters

```{r}
fecundity <- allWing %>%
  group_by(Block, Class, Site_ID) %>%
  dplyr::summarise(wingL=mean(mm, na.rm=T)) %>%
  ungroup() %>%
  mutate(fecundity=-121.240 + (78.02*wingL))
#EFD will be calculated later by dividing multiplying by bite rate
fecundityMean <- mean(fecundity$fecundity)

bc <- infInd %>%
  group_by(Block, Class, Site_ID) %>%
  summarise(bc = mean(Saliva, na.rm = T)) %>%
  ungroup()

bcMean <- mean(bc$bc)
```

## Combine Into VC

Experimental

```{r}
VC <- parameters %>%
  left_join(pEA, by = c("Block", "Class", "Site_ID")) %>%
  left_join(MDR, by = c("Block", "Class", "Site_ID")) %>%
  left_join(fecundity, by = c("Block", "Class", "Site_ID")) %>%
  left_join(bc, by = c("Block", "Class", "Site_ID")) %>%
  #calculate EFD
  mutate(EFD = fecundity * a) %>%
  mutate(VC = ((a^2)*bc*(exp(-mu/PDR))*EFD*pEA*(MDR^2))/((mu^2))) %>%
  mutate(Calculation = "FieldBased")
```

Grand Mean. In this case, the grand mean takes the mean over the parameters effected by carry-over effects, namely fecundity and vector competence.

```{r}
VCmean <- VC %>%
  mutate(bc = bcMean, fecundity = fecundityMean) %>%
  mutate(EFD = fecundity * a) %>%
  mutate(VC = ((a^2)*bc*(exp(-mu/PDR))*EFD*pEA*(MDR^2))/((mu^2))) %>%
  mutate(Calculation = "GrandMean") 
```

Combine Into One Dataframe

```{r}
VCAll <- bind_rows(VC, VCmean)
```

## Save

```{r}
saveRDS(VCAll, "denguePaper/data/infections/clean/VecCapacity.RData")
```
