==> denguePaper/analysis/04_climate-analysis.Rmd <==
---
title: "Microclimate Analysis"
author: "Michelle Evans"
date: ""
output: pdf_document
---

This document contains the code for the analysis of microclimate data in 'Carry-over effects of larval microclimate on the transmission potential of a mosquito-borne pathogen' (Evans et al. 2018).

Paper Citation:

Evans MV, Shiau JC, Solano N, Brindley MA, Drake JM, Murdock, CC. 2018. Carry-over effects of larval microclimate on the transmission potential of a mosquito-borne pathogen.  

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = T)
knitr::opts_chunk$set(cache=TRUE)
knitr::opts_chunk$set(message=FALSE)
knitr::opts_chunk$set(warning=FALSE)

library(lme4)
library(multcomp)
library(lsmeans)

library(tidyverse)
```


## Load Climate By Tray

These values are created in the `01_climate-processing.Rmd` file. They calculate the mean daily mean, minimum, maximum and range per tray over both blocks.

```{r}
climTray <- readRDS("denguePaper/data/microclimate/clean/climateTraySummary.RData")

#add class x block variable
climTray$classxBlock <- paste(climTray$Class, climTray$Block)
```


## Temperature:

```{r}
##------------ Mean T
climModMeanT <- lmer(meanT~Class*Block + (1|Site_ID), data=climTray)
plot(climModMeanT)
summary(climModMeanT)
confint(climModMeanT)
car::Anova(climModMeanT)


summary(lsmeans(climModMeanT, ~Block))

#pairwise stats
summary(glht(climModMeanT, linfct = mcp(Class = "Tukey"), test = adjusted("holm")))

summary(glht(climModMeanT, linfct = mcp(Block = "Tukey"), test = adjusted("holm")))

#pairwise stats-interaction
intMod <- lmer(meanT~classxBlock+ (1|Site_ID), data=climTray)
summary(glht(intMod, linfct = mcp(classxBlock = "Tukey"), test = adjusted("holm")))

##------------ Min T

climModMinT <- lmer(minT~Class*Block + (1|Site_ID), data=climTray)
plot(climModMinT)
summary(climModMinT)
confint(climModMinT)
AIC(climModMinT)
car::Anova(climModMinT)

#pairwise stats
summary(glht(climModMinT, linfct = mcp(Class = "Tukey"), test = adjusted("holm")))

summary(glht(climModMinT, linfct = mcp(Block = "Tukey"), test = adjusted("holm")))

#pairwise stats-interaction
intMod <- lmer(minT~classxBlock+(1|Site_ID), data=climTray)
summary(glht(intMod, linfct = mcp(classxBlock = "Tukey"), test = adjusted("holm")))

##------------ Max T
climModMaxT <- lmer(maxT~Class*Block + (1|Site_ID), data=climTray)
plot(climModMaxT)
summary(climModMaxT)
confint(climModMaxT)

car::Anova(climModMaxT)

#pairwise stats
summary(glht(climModMaxT, linfct = mcp(Class = "Tukey"), test = adjusted("holm")))

#pairwise stats
summary(glht(climModMaxT, linfct = mcp(Block = "Tukey"), test = adjusted("holm")))

#pairwise stats-interaction
intMod <- lmer(maxT~classxBlock+(1|Site_ID), data=climTray)
summary(glht(intMod, linfct = mcp(classxBlock = "Tukey"), test = adjusted("holm")))

##------------ DTR
climModDTR <- lmer(meanDTR~Class*Block + (1|Site_ID), data=climTray)
plot(climModDTR)
summary(climModDTR)
confint(climModDTR)

car::Anova(climModDTR)

#pairwise stats
summary(glht(climModDTR, linfct = mcp(Class = "Tukey"), test = adjusted("holm")))

#pairwise stats
summary(glht(climModDTR, linfct = mcp(Block = "Tukey"), test = adjusted("holm")))

#pairwise stats-interaction
intMod <- lmer(meanDTR~classxBlock+ (1|Site_ID), data=climTray)
summary(glht(intMod, linfct = mcp(classxBlock = "Tukey"), test = adjusted("holm")))

```


## Relative Humidity

```{r}
## ------ mean RH

climModMeanRH <- lmer(meanRH~Class*Block + (1|Site_ID), data=climTray)
plot(climModMeanRH)
summary(climModMeanRH)
#confint(climModMeanRH)
car::Anova(climModMeanRH)

#pairwise stats
summary(glht(climModMeanRH, linfct = mcp(Class = "Tukey"), test = adjusted("holm")))

summary(glht(climModMeanRH, linfct = mcp(Block = "Tukey"), test = adjusted("holm")))

#pairwise stats-interaction
intMod <- lmer(meanRH~classxBlock+ (1|Site_ID), data=climTray)
summary(glht(intMod, linfct = mcp(classxBlock = "Tukey"), test = adjusted("holm")))

anova(climModMeanRH, intMod)
#confidence intervals
lsmeans(climModMeanRH, "Block", type ="response")


## ------ min RH
climModMinRH <- lmer(minRH~Class*Block + (1|Site_ID), data=climTray)
plot(climModMinRH)
summary(climModMinRH)
confint(climModMinRH)
AIC(climModMinRH)

anova(climModMinRH)

#pairwise stats
summary(glht(climModMinRH, linfct = mcp(Class = "Tukey"), test = adjusted("holm")))

summary(glht(climModMinRH, linfct = mcp(Block = "Tukey"), test = adjusted("holm")))

#pairwise stats-interaction
intMod <- lmer(minRH~classxBlock+ (1|Site_ID), data=climTray)
summary(glht(intMod, linfct = mcp(classxBlock = "Tukey"), test = adjusted("holm")))

lsmeans(climModMinRH, pairwise~Class|Block)

## ------ max RH

climModMaxRH <- lmer(maxRH~Class*Block + (1|Site_ID), data=climTray)
plot(climModMaxRH)
summary(climModMaxRH)
# confint(climModMaxRH)
 
anova(climModMaxRH)
car::Anova(climModMaxRH)

anova(climModMaxRH)

#pairwise stats
summary(glht(climModMaxRH, linfct = mcp(Class = "Tukey"), test = adjusted("holm")))

#pairwise stats
summary(glht(climModMaxRH, linfct = mcp(Block = "Tukey"), test = adjusted("holm")))

#pairwise stats-interaction
lsmeans(climModMaxRH, pairwise~Class|Block)


## ------ DHR

climModDHR <- lmer(meanDHR~Class*Block + (1|Site_ID), data=climTray)
plot(climModDHR)
summary(climModDHR)
confint(climModDHR)

car::Anova(climModDHR)

#pairwise stats
summary(glht(climModDHR, linfct = mcp(Class = "Tukey"), test = adjusted("holm")))

#pairwise stats
summary(glht(climModDHR, linfct = mcp(Block = "Tukey"), test = adjusted("holm")))

#pairwise stats-interaction
intMod <- lmer(meanDHR~classxBlock+ (1|Site_ID), data=climTray)
summary(glht(intMod, linfct = mcp(classxBlock = "Tukey"), test = adjusted("holm")))
lsmeans(climModDHR, pairwise~Class|Block)
```


## Save Results

```{r}
# get means in sensical way
extractMean <- function(model, variable){
  meanObj <- summary(lsmeans(model, pairwise~Class*Block))$lsmeans
  results <- meanObj %>%
    mutate(text = paste0(round(lsmean,3), " (", round(lower.CL,3), ",", round(upper.CL, 3), ")")) %>%
    mutate(variable = variable) %>%
    mutate(treatment = paste0(Block, ".", Class)) %>%
    dplyr::select(treatment, variable, text) %>%
    spread(treatment, text)
}
```

```{r}
modelList <- list(model = list(climModMinT, climModMeanT, climModMaxT, climModDTR, 
                               climModMinRH, climModMeanRH, climModMaxRH, climModDHR),
                  variable = list("Min. Temperature", "Mean Temperature", "Max. Temperature", "Daily Temperature Range",
                                  "Min. Relative Humidity", "Mean Relative Humidity", "Max. Relative Humidity", "Daily Humidity Range"))
```

```{r}
results <- pmap_dfr(modelList, extractMean)
```

Write to csv to use in latex

```{r}
write.table(results, "denguePaper/results/tables/climateTable.tsv", sep = '\t', row.names = F)
```
