==> denguePaper/analysis/05_statistical-analysis.Rmd <==
---
title: "Code for 'Carry-over effects of larval microclimate on the transmission potential of a mosquito-borne pathogen' (Evans et al.)"
author: "Michelle Evans"
date: ''
output:
  pdf_document: default
  html_document: default
---

This document contains the code for the analysis of mosquito population and infection data in 'Carry-over effects of urban larval environments on the transmission potential of dengue-2 virus' (Evans et al. 2018).

Paper Citation:

Evans MV, Shiau JC, Solano N, Brindley MA, Drake JM, Murdock, CC. 2018.  Carry-over effects of urban larval environments on the transmission potential of dengue-2 virus.  *BMC Parasites and Vectors*.

# Set-Up

*Note: Working directory should be set to project location.*

```{r knitr setup, echo=F}
knitr::opts_chunk$set(echo = T)
knitr::opts_chunk$set(cache=F)
knitr::opts_chunk$set(message=FALSE)
knitr::opts_chunk$set(warning=FALSE)
knitr::opts_chunk$set(eval=T)
```

```{r packages}
library(lme4) #mixed models
library(broom) #easy model comparison
library(MuMIn)
library(car)
library(MASS)
library(caret)
library(multcomp)
library(lsmeans)

library(tidyverse)
```


# Data Loading and Formatting

## Climate

Climate data that corresponds with infection data.

```{r}
augEnvVar <- readRDS("denguePaper/data/microclimate/clean/summerInfectionClimate.RData")
octEnvVar <- readRDS("denguePaper/data/microclimate/clean/fallInfectionClimate.RData")
```

Tray-level climate data in both summer and fall replicates.

```{r}
climateTray <- readRDS("denguePaper/data/microclimate/clean/climateTraySummary.RData")
```


## Infection

```{r}
seasons <- readRDS("denguePaper/data/infections/clean/seasonInfection.RData")
```

Combine infection with climate that is relevant to infection

```{r}
seasonInf <- merge(seasons, rbind(augEnvVar, octEnvVar), by = c("Block", "Site_ID"))
```

## Direct Effects

```{r}
allSurv <- readRDS("denguePaper/data/emergence/clean/individuals.RData")
emergTray <- readRDS("denguePaper/data/emergence/clean/emergenceTray.RData")
```

Calculate tray level survival and link with climate

```{r}
survTray <- allSurv %>%
  filter(Sex=="F") %>%
  group_by(Tray_ID, Site_ID, Class, Block) %>%
  summarise(survival=mean(event, na.rm = T)) %>%
  ungroup()

survClim <- survTray %>%
  left_join(climateTray, by = c("Block", "Tray_ID", "Class", "Site_ID"))

survBinomial <- allSurv %>%
  filter(Sex == "F") %>%
  group_by(Tray_ID, Site_ID, Class, Block) %>%
  summarise(success = sum(event)) %>%
  mutate(failure = 50 - success) %>%
  ungroup() %>%
  left_join(climateTray, by = c("Block", "Tray_ID", "Class", "Site_ID"))
```

Add tray-level climate data to emergence

```{r}
emergClim <- merge(emergTray, climateTray, by=c("Block","Tray_ID", "Class", "Site_ID"))
```

## Indirect Effects

Wing Length

```{r}
allWing <- readRDS("denguePaper/data/emergence/clean/wingLength.RData")
```

Growth Rate

```{r}
growthDF2 <- readRDS("denguePaper/data/emergence/clean/growthRates.RData")
```

```{r}
VecCapacity <- readRDS("denguePaper/data/infections/clean/VecCapacity.RData")
```

Combine with site level temperature data.

```{r}
siteTemp <- climateTray %>%
  group_by(Block, Class, Site_ID) %>%
  summarise(meanT = mean(meanT))

VecCapacityClimate <- left_join(VecCapacity, siteTemp, by = c("Block", "Class", "Site_ID")) %>%
  filter(Calculation == "FieldBased")
```

# Load Everything to Here to Load all Data

```{r}
```


# Infection Dynamics

Calculate summary stats for table in MS

```{r}
infectionSummary <- seasons %>%
  select(-Wing) %>%
  gather(level, positive, Body:Saliva) %>%
  filter(!(is.na(positive))) %>%
  group_by(Block, Class, level) %>%
  summarise(positives = sum(positive), sampled = n())

#write.csv(infectionSummary, "denguePaper/results/tables/infectionSummary.csv", row.names = F)
```


## Class x Season

```{r final seasonal infection stats}
mixModelseasonsBody21 <- lme4::glmer(Body~Class*Block + (1|Site_ID), 
                          data=seasonInf,
                          family=binomial(link="logit"))
summary(mixModelseasonsBody21)
tidy(mixModelseasonsBody21)
car::Anova(mixModelseasonsBody21) #Wald test
anova(mixModelseasonsBody21)
drop1(mixModelseasonsBody21, test="Chisq") #Likelihood ratio test
lsmeans(mixModelseasonsBody21, list(pairwise~Class|Block, pairwise~Block|Class), omit = 3)


mixModelseasonsHead21 <- lme4::glmer(Head~Class*Block  + (1|Site_ID), 
                          data=seasons,
                          family=binomial(link="logit"))
tidy(mixModelseasonsHead21)
car::Anova(mixModelseasonsHead21) #Wald test
#confint(mixModelseasonsHead21) #profiled confidence interval
drop1(mixModelseasonsHead21, test="Chisq")

lsmeans(mixModelseasonsHead21, list(pairwise~Class|Block, pairwise~Block|Class), omit = 3)

mixModelseasonsSaliva21 <- lme4::glmer(Saliva~Class*Block  + (1|Site_ID), 
                          data=seasons,
                          family=binomial(link="logit"))
tidy(mixModelseasonsSaliva21) #z-test
car::Anova((mixModelseasonsSaliva21)) #Wald test
#confint(mixModelseasonsSaliva21) #profiled confidence interval
drop1(mixModelseasonsSaliva21, test="Chisq")
#pairwise comparison
lsmeans(mixModelseasonsaliva21, list(pairwise~Class|Block, pairwise~Block|Class), omit = 3)
```

## Efficiency

```{r efficiency stats}
#model 
mixModelseasonsBodyEff <- lme4::glmer(Body~Class*Block + (1|Site_ID), 
                          data=seasons,
                          family=binomial(link="logit"))
tidy(mixModelseasonsBodyEff)
car::Anova(mixModelseasonsBodyEff)
summary(multcomp::glht(mixModelseasonsBodyEff, linfct = multcomp::mcp(class = "Tukey"), test = multcomp::adjusted("holm")))
summary(multcomp::glht(mixModelseasonsBodyEff, linfct = multcomp::mcp(Block = "Tukey"), test = multcomp::adjusted("holm")))


mixModelseasonsHeadEff <- lme4::glmer(Head~ Block*Class+ (1|Site_ID), 
                          data=seasons[seasons$Body==1,],
                          family=binomial(link="logit"))
tidy(mixModelseasonsHeadEff)
car::Anova(mixModelseasonsHeadEff) #no effect
drop1(mixModelseasonsHead21, test="Chisq")
summary(multcomp::glht(mixModelseasonsHeadEff, linfct = multcomp::mcp(Block = "Tukey"), test = multcomp::adjusted("holm")))
summary(multcomp::glht(mixModelseasonsHeadEff, linfct = multcomp::mcp(Class = "Tukey"), test = multcomp::adjusted("holm")))

mixModelseasonsSalivaEff <- lme4::glmer(Saliva~Class*Block  + (1|Site_ID), 
                          data=seasons[seasons$Head==1,],
                          family=binomial(link="logit"))
tidy(mixModelseasonsSalivaEff)
car::Anova((mixModelseasonsSalivaEff)) #only Block is significant
drop1(mixModelseasonsSaliva21, test="Chisq")
summary(multcomp::glht(mixModelseasonsSalivaEff, linfct = multcomp::mcp(Class = "Tukey"), test = multcomp::adjusted("holm")))
summary(multcomp::glht(mixModelseasonsSalivaEff, linfct = multcomp::mcp(Block = "Tukey"), test = multcomp::adjusted("holm")))
```

## Microclimate

```{r create dataframe of site-aggregated infections}
seasonSite <- seasons %>%
  dplyr::select(-Individual, - Wing) %>%
  dplyr::group_by(Block, Class, Site_ID) %>%
  summarise_all(funs(mean(.,na.rm=T), se=(sd(., na.rm=T)/sqrt(n())))) %>%
  ungroup()

#group with temperature data
seasonInfSite <- merge(seasonSite, rbind(augEnvVar,octEnvVar), by=c("Block", "Site_ID"))
```

Body by Temp within a season

```{r body x climate stats}
bodySummer <- glm(Body_mean ~ Temp_mean_mean, 
          data=seasonInfSite[seasonInfSite$Block == "Summer", ], 
          family=gaussian(link="logit"))

plot(bodySummer)
car::Anova(bodySummer)
confint(bodySummer)
tidy(bodySummer)
drop1(bodySummer, test="F")


bodyFall <- glm(Body_mean ~ Temp_mean_mean, 
          data=seasonInfSite[seasonInfSite$Block == "Fall", ], 
          family=gaussian(link="logit"))

plot(bodyFall)
car::Anova(bodyFall)
confint(bodyFall)
tidy(bodyFall)
drop1(bodyFall, test="F")
```

Fit Temperature to residuals

```{r}
seasonInf$bodyResid <- NA
seasonInf$bodyResid[!is.na(seasonInf$Body)] <- resid(mixModelseasonsBody21)
bodyResidTemp <- lmer(bodyResid~Temp_mean_mean + (1|Site_ID), 
                          data=seasonInf)
anova(bodyResidTemp)
summary(bodyResidTemp)
tidy(bodyResidTemp)
```

Head by Temp in each season

```{r head x climate stats}
headSummer <- glm(Head_mean ~ Temp_mean_mean, 
          data=seasonInfSite[seasonInfSite$Block == "Summer", ], 
          family=gaussian(link="logit"))

plot(headSummer)
car::Anova(headSummer)
confint(headSummer)
tidy(headSummer)
drop1(headSummer, test="F")


headFall <- glm(Head_mean ~ Temp_mean_mean, 
          data=seasonInfSite[seasonInfSite$Block == "Fall", ], 
          family=gaussian(link="logit"))

plot(headFall)
car::Anova(headFall)
confint(headFall)
tidy(headFall)
drop1(headFall, test="F")
```

Fitting to Residuals:

```{r}
seasonInf$headResid <- NA
seasonInf$headResid[!is.na(seasonInf$Head)] <- resid(mixModelseasonsHead21)
headResidTemp <- lmer(headResid~Temp_mean_mean + (1|Site_ID), 
                          data=seasonInf)
anova(headResidTemp)
plot(headResidTemp)
```

Temperature witin a season:
```{r saliva x climate stats}
salivaSummer <- glm(Saliva_mean+0.01 ~ Temp_mean_mean, 
          data=seasonInfSite[seasonInfSite$Block == "Summer", ],
          family=gaussian(link="logit"))

plot(salivaSummer)
car::Anova(salivaSummer)
confint(salivaSummer)
tidy(salivaSummer)
drop1(salivaSummer, test="F")


salivaFall <- glm(Saliva_mean+0.01 ~ Temp_mean_mean, 
          data=na.omit(seasonInfSite[seasonInfSite$Block == "Fall", ]),
          family=gaussian(link="logit"))

plot(salivaFall)
car::Anova(salivaFall)
confint(salivaFall)
tidy(salivaFall)
drop1(salivaFall, test="F")
```

Fitting to residuals:

```{r}
seasonInf$salivaResid <- NA
seasonInf$salivaResid[!is.na(seasonInf$Saliva)] <- resid(mixModelseasonsSaliva21)
salivaResidTemp <- lmer(salivaResid~Temp_mean_mean + (1|Site_ID), 
                          data=seasonInf)
anova(salivaResidTemp)
plot(salivaResidTemp)
```

## Body Size

### Body Infection

```{r body inf x size stats}
bodyWing <- glmer(Body~Wing + (1|Site_ID),
                          data=seasonInf,
                          family=binomial(link="logit"))

tidy(bodyWing)
summary(bodyWing)

Anova(bodyWing)
drop1(bodyWing, test="Chisq")
```

### Head Infection

```{r head inf x size stats}
headWing <- glmer(Head~Wing + (1|Site_ID),
                          data=seasonInf,
                          family=binomial(link="logit"))

tidy(headWing)
plot(headWing)

Anova(headWing)
```

### Saliva Infection

```{r saliva inf x size stats}
salWing <- glmer(Saliva~Wing + (1|Site_ID),
                          data=seasonInf,
                          family=binomial(link="logit"))
summary(salWing)
tidy(salWing)

Anova(salWing)
```

# Direct Effects

## Survival

### Class x Season

```{r}
SurvModSeason <- glmer(cbind(success, failure) ~ Class*Block + (1|Site_ID),
                   data=survBinomial,
                 family=binomial(link="logit"))

summary(SurvModSeason)
modResults <- tidy(SurvModSeason)
modResults
plot(SurvModSeason)
car::Anova(SurvModSeason)

summary(multcomp::glht(SurvModSeason, linfct = multcomp::mcp(Block = "Tukey"), test = adjusted("holm")))

interaction.plot(survTray$Class, survTray$Block, survTray$survival)

#confidence intervals
lsmeans(SurvModSeason, "Block", type ="response")
```

### Microclimate

Temperature within a season:

```{r}
survModClimSummer <- glmer(cbind(success, failure)~meanT+(1|Site_ID),
                    data=survBinomial[survBinomial$Block=="Summer",],
                    family=binomial("logit"))

summary(survModClimSummer)
plot(survModClimSummer)
qqnorm(resid(survModClimSummer))

survModClimFall <- glmer(cbind(success, failure)~meanT+(1|Site_ID),
                    data=survBinomial[survBinomial$Block=="Fall",],
                    family=binomial("logit"))

summary(survModClimFall)
plot(survModClimFall)
qqnorm(resid(survModClimFall))
```

Fitting to residuals:

```{r}
survBinomial$survResid <- NA
survBinomial$survResid <- resid(SurvModSeason)
survResidTemp <- lmer(survResid~meanT + (1|Site_ID), 
                          data=survBinomial)
anova(survResidTemp)
plot(survResidTemp)
car::Anova(survResidTemp)
```

## Emergence

### Class x Season

```{r}
#do not include trays with no survival as devRate could not be estimated (==0)
mixEmergeTray <- lmerTest::lmer(devRate ~ Class*Block +(1|Site_ID), 
                   data=emergTray[emergTray$devRate!=0,])

summary(mixEmergeTray)
tidy(mixEmergeTray)
plot(mixEmergeTray)
qqnorm(resid(mixEmergeTray))
#confint(mixEmerge)
Anova(mixEmergeTray)

#pairwise
summary(multcomp::glht(mixEmergeTray, linfct = multcomp::mcp(Block = "Tukey"), test = adjusted("holm")))
```

### Microclimate

Within a season:
```{r}
devClimModSummer <- lmerTest::lmer(devRate~meanT + (1|Site_ID),
                    data=emergClim[emergClim$devRate!=0 & emergClim$Block=="Summer",])

plot(devClimModSummer)
qqnorm(resid(devClimModSummer))
summary(devClimModSummer)

devClimModFall <- lmerTest::lmer(devRate~meanT + (1|Site_ID),
                    data=emergClim[emergClim$devRate!=0 & emergClim$Block=="Fall",])

plot(devClimModFall)
qqnorm(resid(devClimModFall))
summary(devClimModFall)
```

Fitting residuals:

```{r}
emergClim$emergResid <- NA
emergClim$emergResid[emergClim$devRate!=0] <- resid(mixEmergeTray)
emergResidTemp <- lmer(emergResid~meanT + (1|Site_ID), 
                          data=emergClim[emergClim$devRate!=0,])
anova(emergResidTemp)
plot(emergResidTemp)
car::Anova(emergResidTemp)
```

# Indirect Effects

## Body Size (Uninfected)

### Land Class x Season

```{r}
wingMod <- lmerTest::lmer(mm~Class*Block +(1|Site_ID),
                data=allWing)
summary(wingMod)
plot(wingMod)
car::Anova(wingMod)

lsmeans(wingMod, pairwise~Block|Class)

interaction.plot(allWing$Class, allWing$Block, allWing$mm)
```

### Microclimate

Statistics:
```{r}
wingModTempSummer <- lmerTest::lmer(mm~Temp+(1|Site_ID),
                    data=allWing[allWing$Block=="Summer", ])
plot(wingModTempSummer)
summary(wingModTempSummer)
#confint(wingModTemp) #no effect of temperature
qqnorm(resid(wingModTempSummer))
car::Anova(wingModTempSummer)

wingModTempFall <- lmerTest::lmer(mm~Temp+(1|Site_ID),
                    data=allWing[allWing$Block=="Fall", ])
plot(wingModTempFall)
summary(wingModTempFall)
#confint(wingModTemp) #no effect of temperature
qqnorm(resid(wingModTempFall))
car::Anova(wingModTempFall)
```

Fitting Residuals:

```{r}
allWing$wingResid <- NA
allWing$wingResid<- resid(wingMod)
wingResidTemp <- lmer(wingResid~Temp + (1|Site_ID), 
                          data=allWing)
anova(wingResidTemp )
plot(wingResidTemp )
car::Anova(wingResidTemp)
```

## Growth Rate

### Class x Season

```{r}
growthModSeason <- lmer(r ~ Class*Block+ (1|Site_ID), 
                        data=growthDF2)

summary(growthModSeason)
plot(growthModSeason)

tidy(growthModSeason)
car::Anova(growthModSeason) #Wald test
#confint(growthModSeason) #profiled confidence interval
```

### Microclimate

Within a season:
```{r}
growthTempModSummer <- lmerTest::lmer(r~meanT + (1|Site_ID),
                      data=growthDF2[growthDF2$Block=="Summer",])
summary(growthTempModSummer)
anova(growthTempMod)
plot(growthTempMod)

growthTempModFall<- lmerTest::lmer(r~meanT + (1|Site_ID),
                      data=growthDF2[growthDF2$Block=="Fall",])
summary(growthTempModFall)
anova(growthTempModFall)
```

Fitting temperature to residuals:
```{r}
growthDF2$r_resid <- NA
growthDF2$r_resid <- resid(growthModSeason)
growthResidTemp<- lmer(r_resid~ meanT + (1|Site_ID), 
                          data=growthDF2)
anova(growthResidTemp)
```

# Vectorial Capacity

## Class x Season

Is the grand mean different than field derived? We will look at this qualitatively in the plot, and maybe use some coef variation terms.

```{r}
VCModelBoth <- lmer(VC ~ Class * Block * Calculation + (1|Site_ID),
                     data = VecCapacity) 

summary(VCModelBoth)
tidy(VCModelBoth)
Anova(VCModelBoth)

VCpredsBoth <- cbind(VecCapacity, predict(VCModelBoth, response = T))
```

Because there is no significant difference, I only report the other stats on the field-based calculation.

```{r}
VCModel <- lm(VC~ Class * Block,
                     data = VecCapacityClimate) 

plot(VCModel)
summary(VCModel)
car::Anova(VCModel)
tidy(VCModel)

lsmeans(VCModel, pairwise ~ Class|Block)

#confidence intervals
lsmeans(VCModel, "Block", type ="response")
```

Using `anova` for comparison:
```{r}
#is it balanced? yes
table(VecCapacityClimate$Block, VecCapacityClimate$Class)

VC.aov <- aov(VC~ Class * Block,
                     data = VecCapacityClimate)
summary(VC.aov, type = 3)
anova(VCModel)
#they are the same
```


## Microclimate

Within a season:
```{r}
VCModSummer <- lm(VC~ meanT,
                     data = VecCapacityClimate[VecCapacityClimate$Block=="Summer",]) 
summary(VCModSummer)
anova(VCModSummer)
plot(VCModSummer)

VCModFall <- lm(VC~ meanT,
                     data = VecCapacityClimate[VecCapacityClimate$Block=="Fall",]) 
summary(VCModFall)
anova(VCModFall)
plot(VCModFall)
```

Fitting residuals:

```{r}
VecCapacityClimate$vc_resid <- NA
VecCapacityClimate$vc_resid <- resid(VCModel)
vcResidTemp <- lm(vc_resid~ meanT, 
                          data=VecCapacityClimate)
anova(vcResidTemp)
plot(vcResidTemp)
car::Anova(vcResidTemp)
```

# Supplementary Tables

## Temperature within a season

```{r}
seasonModelsSummer <- list(survModClimSummer,  devClimModSummer, wingModTempSummer, growthTempModSummer, bodySummer, headSummer, salivaSummer, VCModSummer)
names(seasonModelsSummer) <- c("Survival", "Development", "Wing Length", "Growth", "Body Inf", "Head Inf", "Sal Inf", "VC")

summerFixefs <- plyr::ldply(list(survModClimSummer,  devClimModSummer, wingModTempSummer, growthTempModSummer), fixef)[,2]
summerCoefs <- plyr::ldply(list(bodySummer, headSummer, salivaSummer, VCModSummer), coef)[,2]

summerStats <- plyr::ldply(seasonModelsSummer, Anova, .id = "model")[-9,]
summerStats$Season <-"Summer"
summerStats$coef <- c(summerFixefs, summerCoefs)
```

```{r}
#for ANOVA type II
seasonModelsFall <- list(survModClimFall,  devClimModFall, wingModTempFall, growthTempModFall, bodyFall, headFall, salivaFall, VCModFall)
names(seasonModelsFall) <- c("Survival", "Development", "Wing Length", "Growth", "Body Inf", "Head Inf", "Sal Inf", "VC")

#for coefficients
fallFixefs <- plyr::ldply(list(survModClimFall,  devClimModFall, wingModTempFall, growthTempModFall), fixef)[,2]
fallCoefs <- plyr::ldply(list(bodyFall, headFall, salivaFall, VCModFall), coef)[,2]

fallStats <- plyr::ldply(seasonModelsFall, Anova, .id = "model")[-9,]
fallStats$Season <- "Fall"
fallStats$coef <- c(fallFixefs, fallCoefs)
```

Combine and save

```{r}
seasonalTempStats <- rbind(summerStats,fallStats)[,c(9,1:8, 10)]

write.csv(seasonalTempStats, "denguePaper/results/tables/temperature-withinSeason.csv", row.names = F)
```

## Fitting Temperature to Residuals

```{r}
#put models into a list to tidy up
residModels <- list(survResidTemp, emergResidTemp, wingResidTemp, growthResidTemp, bodyResidTemp, headResidTemp, salivaResidTemp, vcResidTemp)
names(residModels) <- c("Survival", "Development", "Wing Length", "Growth", "Body Inf", "Head Inf", "Sal Inf", "VC")

residualAnova <- plyr::ldply(residModels, anova, .id = "model")[-9,-c(3,6)] 
```

```{r}
write.csv(residualAnova, "denguePaper/results/tables/temperature-residuals.csv", row.names = F)
```
