==> denguePaper/analysis/06_figure-creation.Rmd <==
---
title: "Figures"
author: "Michelle Evans"
date: "March 21, 2018"
output: html_document
---

This document makes the figures for 'Carry-over effects of larval microclimate on the transmission potential of a mosquito-borne pathogen' (Evans et al.).

Paper Citation:

Evans MV, Shiau JC, Solano N, Brindley MA, Drake JM, Murdock, CC. Carry-over effects of larval microclimate on the transmission potential of a mosquito-borne pathogen. 

Data is loaded individually for each figure so they are easier to make individually.


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#packages
library(ggthemes)
library(raster)
library(sp)
library(ggmap)
library(rgdal)
library(rasterVis)
library(latticeExtra)
library(scales)
library(png)
library(grid)
library(cowplot)

library(tidyverse)

```


```{r set colors}
colR <- "dodgerblue"
colS <- "gray60"
colU <- "maroon"

axisColor <- "gray20"
errorColor <- "gray20"
```

# Microclimate

```{r}
climate <- readRDS("denguePaper/data/microclimate/clean/2016TrialsAdult.RData")
```

## Summary Metrics

Calculate the mean minima, maximia, and mean by class by day from 2016-08-01 to 2016-11-07. This is to plot in figures over the whole time period.

```{r}
climateSummary <- climate %>%
  dplyr::group_by(Tray_ID, Day, Site_ID, Class) %>%
  dplyr::summarise(meanT=mean(Temp), minT=min(Temp), maxT=max(Temp), meanRH=mean(RH), maxRH=max(RH), minRH=min(RH)) %>% #values per tray
  ungroup() %>%
  group_by(Class, Day) %>% #get mean values per class
  dplyr::summarise(meanT=mean(meanT), minT=mean(minT), maxT=mean(maxT), meanRH=mean(meanRH), maxRH=mean(maxRH), minRH=mean(minRH)) %>%
  ungroup()
```

## Temperature Plot:

```{r, eval = F}
temperature <- ggplot(data=climateSummary, aes(x=Day, group=Class))+
  geom_line(aes(y=meanT, col=Class)) + 
  geom_line(aes(y=minT, col=Class), linetype="dotted") +
  geom_line(aes(y=maxT, col=Class), linetype="dotted") +
  theme_fivethirtyeight() +
  scale_color_manual(values=c("dodgerblue", "gray20", "maroon")) +
  theme(axis.title = element_text(), axis.title.x = element_blank()) +
  ylab("Temperature (C)") + 
  annotate("rect", xmin=as.Date("2016-08-01", format="%Y-%m-%d"), xmax=as.Date("2016-09-01", format="%Y-%m-%d"), ymin=-Inf, ymax=Inf, fill="gray20", alpha=0.1)+
    annotate("rect", xmin=as.Date("2016-09-26", format="%Y-%m-%d"), xmax=as.Date("2016-11-07", format="%Y-%m-%d"), ymin=-Inf, ymax=Inf, fill="gray20", alpha=0.1)+
  annotate("text", x = as.Date("2016-08-15", format="%Y-%m-%d"), y = 3, label = "Summer") +
  annotate("text", x = as.Date("2016-10-18", format="%Y-%m-%d"), y = 3, label = "Fall") +
  theme(panel.background = element_rect(fill = "transparent", colour = NA), 
        plot.background = element_rect(fill = "transparent", colour = NA),
        legend.key = element_blank(),
        axis.line=element_line(color=ggthemes_data$fivethirtyeight["dkgray"], size=0.5),
        panel.grid = element_blank())+
  theme(legend.background = element_rect(fill = "transparent", colour = NA))

```

## Relative Humidity

```{r, eval=F}
rh <- ggplot(data=climateSummary, aes(x=Day, group=Class))+
  geom_line(aes(y=meanRH, col=Class)) + 
  geom_line(aes(y=minRH, col=Class), linetype="dotted") +
  geom_line(aes(y=maxRH, col=Class), linetype="dotted") +
  theme_fivethirtyeight() +
  scale_color_manual(values=c("dodgerblue", "gray20", "maroon")) +
  theme(axis.title = element_text(), axis.title.x = element_blank()) +
  ylab("Relative Humidity (%)") + 
  xlab("Date")+
  annotate("rect", xmin=as.Date("2016-08-01", format="%Y-%m-%d"), xmax=as.Date("2016-09-01", format="%Y-%m-%d"), ymin=-Inf, ymax=Inf, fill="gray20", alpha=0.1)+
    annotate("rect", xmin=as.Date("2016-09-26", format="%Y-%m-%d"), xmax=as.Date("2016-11-07", format="%Y-%m-%d"), ymin=-Inf, ymax=Inf, fill="gray20", alpha=0.1)+
    annotate("text", x = as.Date("2016-08-15", format="%Y-%m-%d"), y = 22, label = "Summer") +
  annotate("text", x = as.Date("2016-10-18", format="%Y-%m-%d"), y = 22, label = "Fall") +
  theme(panel.background = element_rect(fill = "transparent", colour = NA), 
        plot.background = element_rect(fill = "transparent", colour = NA),
        legend.key = element_blank(),
        axis.line=element_line(color=ggthemes_data$fivethirtyeight["dkgray"], size=0.5),
        panel.grid = element_blank())+
  theme(legend.background = element_rect(fill = "transparent", colour = NA))
#dev.off()
```

## Create Plot

```{r}
pdf(file="denguePaper/results/figures/forMS/climate.pdf", width = 6, height=8, family="sans")
plot_grid(temperature, rh, align="v", nrow=2)
dev.off()
```


# Population Dynamics

```{r}
survTray <-  readRDS("denguePaper/data/emergence/clean/survivalTray.RData")
emergTray <- readRDS("denguePaper/data/emergence/clean/emergenceTray.RData")
growthDF2 <- readRDS("denguePaper/data/emergence/clean/growthRates.RData")
```


```{r}
allDemos <- growthDF2 %>%
  dplyr::select(Block, Class, Site_ID, Tray_ID, r) %>%
  left_join(emergTray, by = c("Block", "Class", "Site_ID", "Tray_ID")) %>%
  left_join(survTray, by =c("Block", "Class", "Site_ID", "Tray_ID")) %>%
  gather(trait, value, r:numFSurv)

allDemoSite <- allDemos %>%
  select(-Tray_ID) %>%
  group_by(Block, Class, Site_ID, trait) %>%
  summarise_all(funs(mean = mean, se = sd(.)/n())) %>%
  ungroup() 
```

Plot

```{r}
#pdf(file="denguePaper/results/figures/forMS/demographicPoints.pdf", width = 3.25, height=6, family="sans")

tiff(file="denguePaper/results/figures/forMS/demographicPoints.tiff", width = 975, height = 1800, family="sans", res = 300)

demographicPlot <- ggplot(data = allDemoSite, aes(x = as.factor(Class), y = mean)) +
  geom_boxplot(aes(alpha = Block), color = NA) +
  geom_pointrange(aes(fill = Class, shape = Block, group = Block, ymin = mean-se, ymax = mean+se), 
                  color = "black", 
                  position = position_dodge(width = 0.75), 
                  size = 0.5, stroke = 0.5, lwd = 0.5) +
  facet_wrap(~trait, dir = "v", scales = "free_y", strip.position = "left",
             labeller = as_labeller(c(devRate = "Development Rate\n(1/day)", numFSurv = "Egg-to-Adult\nSurvival (%)", r ="Per Capita\nGrowth Rate (r')"))) +
  guides(shape = F, alpha = F, color = F, fill = F) +
  xlab("Land Class") +
  scale_color_manual(values=c(colR, colS, colU), labels=c("Rural", "Suburban", "Urban")) +
  scale_fill_manual(values=c(colR, colS, colU), labels=c("Rural", "Suburban", "Urban")) +
  scale_shape_manual(values = c(21, 23)) +
  theme_base() +
  theme(
        strip.background = element_blank(),
        strip.text = element_text(size=11),
        strip.placement = "outside",
        axis.title.y = element_blank(),
        axis.text = element_text(size = 10),
        axis.title.x = element_text(size = 11)
  )

ggdraw()+
  draw_plot(demographicPlot, x=0.0, y=0.0, width=1, height=1)+
  draw_plot_label(label=c("a", "b", "c"),
                  x=c(0,0,0),
                  y=c(1,0.71,0.40))

dev.off()
  
```

# Infection Dynamics

## Class x Block

```{r infection data format}
infLong <- readRDS("denguePaper/data/infections/clean/infectionSummaryLong.RData")
```

```{r}
pdf(file="denguePaper/results/figures/forMS/infectionFacet2.pdf", width = 3.25, height=6, family="sans")

infectionPlot <- ggplot(data = infLong, aes(x=Class, group=Block))+
  geom_bar(stat="identity", aes(y=mean.inf, alpha=Block, fill=Class), color="gray20", position=position_dodge(width=0.9))+
  scale_y_continuous(breaks=c(0,0.5,1), minor_breaks = c(0.25, 0.75), labels=c("0%", "50%", "100%"), limits=c(0,1))+
  scale_fill_manual(values=c(colR, colS, colU))+
  scale_alpha_discrete(range=c(1,0.1), name="Season")+
  geom_errorbar(aes(ymin=mean.inf-se.inf, ymax=mean.inf+se.inf), width=0.2, color=errorColor, position = position_dodge(0.9))+
  xlab("Land Class") +
  ylab("Percent Positive")+
  guides(fill = F, color = F,
         alpha = F) +
  theme_base() +
  facet_wrap(~type, dir = "v") +
  theme(
        strip.background = element_blank(),
        strip.text = element_blank(),
        axis.text = element_text(size = 10),
        axis.title = element_text(size = 11)
)

ggdraw()+
  draw_plot(infectionPlot, x=0.0, y=0.0, width=1, height=1)+
  draw_plot_label(label=c("a", "b", "c"),
                  x=c(0,0,0),
                  y=c(1,0.71,0.40))

dev.off()


```
## Temperature

```{r}
seasons <- readRDS("denguePaper/data/infections/clean/seasonInfection.RData")
augEnvVar <- readRDS("denguePaper/data/microclimate/clean/summerInfectionClimate.RData")
octEnvVar <- readRDS("denguePaper/data/microclimate/clean/fallInfectionClimate.RData")

seasonSite <- seasons %>%
  dplyr::select(-Individual, - Wing) %>%
  dplyr::group_by(Block, Class, Site_ID) %>%
  summarise_all(funs(mean(.,na.rm=T), se=(sd(., na.rm=T)/sqrt(n())))) %>%
  ungroup()

#group with temperature data
seasonInfSite <- merge(seasonSite, rbind(augEnvVar,octEnvVar), by=c("Block", "Site_ID"))
```


```{r}
pdf(file="denguePaper/results/figures/extra/InfxTemp.pdf", width = 8, height=3, family="sans")

plotBody <- ggplot(data=seasonInfSite, aes(x=Temp_mean_mean, y=Body_mean))+
  geom_errorbar(aes(ymin=Body_mean-Body_se, 
                    ymax=Body_mean+Body_se), color=errorColor, width=0.1) +
  geom_point(aes(color=Class, shape=Block), size=4.5) + 
  scale_color_manual(values=c(colR, colS, colU)) +
  theme_bw() +
  theme(panel.background = element_rect(fill = "transparent", colour = NA),
        plot.background = element_rect(fill = "transparent", colour = NA),
        legend.key = element_blank(),
        legend.background = element_rect(fill = "transparent", colour = NA),
        axis.title = element_text(),
        axis.title.x = element_text(),
        axis.title.y=element_text(),
        axis.line=element_line(color=axisColor, size=0.5),
        panel.grid = element_blank())+
  ylab("Prop. Infected")+
  xlab("") +  
  theme(legend.position="none")

plotHead <- ggplot(data=seasonInfSite, aes(x=Temp_mean_mean, y=Head_mean))+
   geom_errorbar(aes(ymin=Head_mean-Head_se, 
                     ymax=Head_mean+Head_se), color=errorColor, width=0.1) +
  geom_point(aes(color=Class, shape=Block), size=4) + 
  scale_color_manual(values=c(colR, colS, colU)) +
  theme_bw() +
    theme(panel.background = element_rect(fill = "transparent", colour = NA),
        plot.background = element_rect(fill = "transparent", colour = NA),
        legend.key = element_blank(),
        legend.background = element_rect(fill = "transparent", colour = NA),
        axis.title = element_text(),
        axis.title.x = element_text(),
        axis.title.y=element_blank(),
        axis.line=element_line(color=axisColor, size=0.5),
        panel.grid = element_blank())+
  xlab("Mean Daily Temperature (C)")+
  theme(legend.position="none")

plotSaliva <- ggplot(data=seasonInfSite, aes(x=Temp_mean_mean, y=Saliva_mean))+
  geom_errorbar(aes(ymin=Saliva_mean-Saliva_se, ymax=Saliva_mean+Saliva_se), color=errorColor, width=0.1) +
  geom_point(aes(color=Class, shape=Block), size=4.5) + 
  scale_color_manual(values=c(colR, colS, colU)) +
  theme_bw() +
    theme(panel.background = element_rect(fill = "transparent", colour = NA),
        plot.background = element_rect(fill = "transparent", colour = NA),
        legend.key = element_blank(),
        legend.background = element_rect(fill = "transparent", colour = NA),
        axis.title = element_text(),
        axis.title.x = element_text(),
        axis.title.y=element_blank(),
        axis.line=element_line(color=axisColor, size=0.5),
        panel.grid = element_blank()) +
  xlab("") +  
  scale_y_continuous(limits=c(0,1))+
  theme(legend.position="none")


plot_grid(plotBody, plotHead, plotSaliva, 
          labels=c("A", "B", "C"), 
          ncol=3)

dev.off()
```

## Wing Length

```{r}
seasons <- readRDS("denguePaper/data/infections/clean/seasonInfection.RData")
augEnvVar <- readRDS("denguePaper/data/microclimate/clean/summerInfectionClimate.RData")
octEnvVar <- readRDS("denguePaper/data/microclimate/clean/fallInfectionClimate.RData")
seasonInf <- seasonInf <- merge(seasons, rbind(augEnvVar, octEnvVar), by = c("Block", "Site_ID"))
```

```{r body inf x size plot, eval=F, include=F}
ggplot(data=seasonInf[!is.na(seasonInf$Body),], aes(x=factor(Body), y=Wing))+
  geom_boxplot() +
  geom_jitter(shape=16, position=position_jitter(0.1), aes(color=factor(Class)), alpha=0.5) +
  theme_fivethirtyeight()+
  theme(axis.title = element_text()) +
  theme(panel.background = element_rect(fill = "transparent", colour = NA), 
        plot.background = element_rect(fill = "transparent", colour = NA),
        legend.key = element_blank())+
  theme(legend.background = element_rect(fill = "transparent", colour = NA))+
  theme(legend.position="right", legend.direction="vertical") +
  scale_color_manual(values=c("dodgerblue", "gray10", "maroon"), labels=c("Rural", "Suburban", "Urban"))+
    scale_fill_manual(values=c("dodgerblue", "gray80", "maroon"), labels=c("Rural", "Suburban", "Urban")) +
  guides(color=F) +
  xlab("Body Positivity") +
  ylab("Wing Size")
```

# Vectorial Capacity

```{r}
VecCapacity <- readRDS("denguePaper/data/infections/clean/VecCapacity.RData")
climateTray <- readRDS("denguePaper/data/microclimate/clean/climateTraySummary.RData")
```

```{r}
siteTemp <- climateTray %>%
  group_by(Block, Class, Site_ID) %>%
  summarise(meanT = mean(meanT))

VecCapacityClimate <- left_join(VecCapacity, siteTemp, by = c("Block", "Class", "Site_ID"))
```

Plot across temperature with the mean value per block
```{r}
VecCapacityBlock <- VecCapacityClimate %>%
  dplyr::select(Block, Calculation, VC, meanT) %>%
  group_by(Block, Calculation) %>%
  summarise(VC.mean = mean(VC), VC.sd = sd(VC), meanT = mean (meanT), VC.median = median(VC)) %>%
  ungroup() %>%
  #offset to plot next to each other
  mutate(offset = case_when(
    Calculation == "FieldBased" ~ meanT - 0.5,
    Calculation == "GrandMean" ~ meanT + 0.5
  ))
```


```{r}
range.box <- function(x, y.mean, y.sd, id, Calculation) {
  data.frame(
    id = id,
    x = rep(c(x - 0.45, x + 0.45),each = 2),
    y = c(y.mean - y.sd, y.mean + y.sd, y.mean + y.sd, y.mean - y.sd),
    Calculation = Calculation
  )
}
```

```{r}
data2purr <- list(x = VecCapacityBlock$offset, 
                  y.mean =  VecCapacityBlock$VC.mean,
                  y.sd = VecCapacityBlock$VC.sd,
                  id = 1:4,
                  Calculation = VecCapacityBlock$Calculation)
comparePolygons <- pmap_dfr(data2purr, range.box)
rm(data2purr)
```


```{r}
pdf(file="denguePaper/results/figures/forMS/VC.pdf", width = 4, height=4, family="sans")

tiff(file = "denguePaper/results/figures/forMS/VC.tiff", width = 1200, height = 1200, res = 300, family = "sans")

ggplot(data = VecCapacityClimate) +
  geom_polygon(data=comparePolygons, aes(x = x, y = y, group = id),
               linetype = rep(c("dotted", "solid", "dotted", "solid"), each = 4),
               fill = NA,
               color = "black") +
  geom_errorbar(data = VecCapacityBlock, aes(x=offset, ymin = VC.median, ymax = VC.median), color = "black")+
  geom_point(data = VecCapacityClimate, aes(x=meanT, y = VC, fill = Class, shape = Calculation), 
             size = 3, color = "black") +
  theme_base() +
  ylab("Vectorial Capacity") +
  xlab("Mean Temperature (C)") +
  scale_fill_manual(values = c(colR, colS, colU), 
                    labels = c("Rural", "Suburban", "Urban"),
                    guide = "legend") +
  scale_shape_manual(values = c(21, 23)) +
  theme(legend.position = c(0.2,0.8),
        legend.text = element_text(size=6),
        legend.title = element_text(size=8),
        legend.key.size = unit(1, "line")) +
  guides(fill = guide_legend(override.aes = list(shape = 21)),
         shape = "none")

dev.off()
```

Plot of individual points ( VC by site). Box plots represent median VC values per block and calculation types, with box range representing mean +_ sd.

## Plot of Bias

```{r}
VecCapacityDiff <- VecCapacity %>%
  dplyr::select(Block, Class, Site_ID, VC, Calculation) %>%
  spread(Calculation, VC) %>%
  mutate(DeltaVC = GrandMean - FieldBased) %>%
  group_by(Block, Class) %>%
  summarise(meanDiff = mean(DeltaVC), seDiff = sd(DeltaVC)/sqrt(3))
```


```{r}
pdf(file="denguePaper/results/figures/forMS/SuppVCDiff.pdf", width = 4, height=4, family="sans")

ggplot(data = VecCapacityDiff, aes(x = Class, y = meanDiff, group = Block)) +
  geom_bar(stat = 'identity', aes(fill = Class, alpha = Block), position=position_dodge(width=0.9), color = "black") +
  scale_fill_manual(values=c(colR, colS, colU))+
  scale_color_manual(values=c(colR, colS, colU))+
  scale_alpha_discrete(range=c(1,0.1), name="Season")+
  ylab("Absolute Change in VC\ndue to Grand Mean") +
  guides(fill = F) +
  theme_bw() +
    theme(panel.background = element_rect(fill = "transparent", colour = NA), 
        plot.background = element_rect(fill = "transparent", colour = NA),
        legend.key = element_blank(),
        panel.grid.major.x = element_blank(),
        legend.background = element_rect(fill = "transparent", colour = NA),
        axis.line=element_line(color=axisColor, size=0.5),
        panel.grid = element_blank(),
        axis.text.y=element_text(size = 10),
        axis.text.x=element_text(size = 10),
        legend.text = element_text(size = 10),
        legend.title = element_text(size = 10))
  
dev.off()
```

Bias introduced by not accounting for carry-over effects. Bar plots represent mean value of the difference between the grand mean calculation and field based calculation of VC across land class and block.
# Map of Study Site

```{r}
impSurf <- raster("denguePaper/data/spatial/impSurfFocal.tif")
sites <- read.csv("denguePaper/data/spatial/sites.csv", header=T)
sitesSp <- SpatialPointsDataFrame(sites[,c('X','Y')], data= sites,
                                proj4string=crs("+init=epsg:4326"))
sitesSp <- spTransform(sitesSp, "+proj=utm +zone=17 +datum=NAD83 +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0")

counties <- readOGR("denguePaper/data/spatial/","cb_2016_us_county_500k") #downloaded from US census bureau 2015
acc <- counties[counties$GEOID==13059,]
acc <- spTransform(acc, "+proj=utm +zone=17 +datum=NAD83 +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0")
```

```{r}
#bound Athens around sites
athens <- crop(impSurf, c(262000, 297000, 3748000,3765000))
```

```{r}
#load map of georgia
gaMap <- readPNG("denguePaper/data/spatial/accMap.png")
mapRaster<- as.raster(gaMap)
```

```{r}
png("denguePaper/results/figures/forMS/siteMap.png", family="sans", width=7, height=4, units="in", res=500)
levelplot(athens, 
          margin=FALSE,                       
          colorkey=list(
            space='bottom',                   
            labels=list(at=c(0,5,40,100), 
                        labels=c("0%","5%", "40%", "100%"),
                        font=4),
            axis.line=list(col='black')       
          ),    
          par.settings=list(
            axis.line=list(col='gray20') 
          ),
          scales=list(draw=FALSE),
          col.regions=c("white", alpha("dodgerblue",0.7), alpha("maroon",0.7)),
          at=c(0,5,40,100)) +           
  layer(sp.polygons(sitesSp, 
                    pch=c(15,16,17)[sitesSp$CLASS], 
                    fill="gray20",
                    cex=1, lwd=2
                    ))+
  layer(sp.polygons(acc, lwd=2, col="black"))+
  layer(SpatialPolygonsRescale(layout.north.arrow(), offset = c(265000, 3749000), scale = 2000)) +
  layer({
    xs <- seq(267000, 269250, by=750)
    grid.rect(x=xs, y=3750000,
              width=750, height=300,
              gp=gpar(fill=rep(c('transparent', 'black'), 2)),
              default.units='native')
    grid.text(x=seq(266800, 270000, by=750), y=rep(c(3749500, 3750400), 2), label=c("","", "1.5", "", "3 km"),
              gp=gpar(cex=0.5), rot=0,
              default.units='native')
  })

dev.off()
```



# Infection Efficicency

```{r plot efficiency, eval=F}
efficiencyPlot <- readRDS("denguePaper/data/infections/clean/efficiencyPlot.RData")

#supplemental plot
ggplot(data=efficiencyPlot, aes(x=Class, group=Block))+
  geom_bar(stat="identity", aes(y=mean, alpha=factor(Block), fill=factor(Class)), color="gray20", position=position_dodge(width=0.9))+
  facet_wrap(~type, nrow=3, dir="v") +
  scale_fill_manual(values=c(colR, colS, colU))+
  scale_alpha_discrete(range=c(1,0.1), name="Season")+
  geom_errorbar(aes(ymin=mean-SE, ymax=mean+SE), width=0.2, color="gray20", position =position_dodge(0.9))+
  theme_base() + 
  xlab("Land Class") +
  ylab("Infection Efficiency") +
  guides(fill=F)
```

# Temperature and Variables

## Demographic

```{r}
survTray <-  readRDS("denguePaper/data/emergence/clean/survivalTray.RData")
emergTray <- readRDS("denguePaper/data/emergence/clean/emergenceTray.RData")
growthDF2 <- readRDS("denguePaper/data/emergence/clean/growthRates.RData")
```

```{r}
pdf(file="denguePaper/results/figures/extra/growthTemp.pdf", width = 7, height=4, family="sans")

ggplot(data=growthDF2, aes(x = meanT, y = r, group = Site_ID)) +
  geom_point(aes(color = Class)) +
  geom_smooth(method = "lm", aes(group = Class, color = Class), se = F) +
  facet_wrap(~Block, scales = "free_x") +
  theme_base()

dev.off()
```


## Infection
